chore: rm crate reth-rpc-types
