+++ crates/rpc/rpc-builder/Cargo.toml
@@ -62,7 +62,6 @@ reth-payload-builder = { workspace = true, features = ["test-utils"] }
 reth-provider = { workspace = true, features = ["test-utils"] }
 reth-rpc-api = { workspace = true, features = ["client"] }
 reth-rpc-engine-api.workspace = true
-reth-rpc-types.workspace = true
 reth-tracing.workspace = true
 reth-transaction-pool = { workspace = true, features = ["test-utils"] }
 reth-tokio-util.workspace = true
