+++ crates/rpc/rpc-eth-types/Cargo.toml
@@ -22,7 +22,6 @@ reth-primitives = { workspace = true, features = ["secp256k1"] }
 reth-storage-api.workspace = true
 reth-revm.workspace = true
 reth-rpc-server-types.workspace = true
-reth-rpc-types.workspace = true
 reth-rpc-types-compat.workspace = true
 reth-tasks.workspace = true
 reth-transaction-pool.workspace = true
