+++ crates/rpc/rpc-eth-types/src/error.rs
@@ -10,7 +10,6 @@ use reth_primitives::{revm_primitives::InvalidHeader, BlockId};
 use reth_rpc_server_types::result::{
     block_id_to_str, internal_rpc_err, invalid_params_rpc_err, rpc_err, rpc_error_with_code,
 };
-use reth_rpc_types::ToRpcError;
 use reth_transaction_pool::error::{
     Eip4844PoolTransactionError, Eip7702PoolTransactionError, InvalidPoolTransactionError,
     PoolError, PoolErrorKind, PoolTransactionError,
@@ -18,6 +17,18 @@ use reth_transaction_pool::error::{
 use revm::primitives::{EVMError, ExecutionResult, HaltReason, InvalidTransaction, OutOfGasError};
 use revm_inspectors::tracing::MuxError;
 use tracing::error;
+
+/// A trait to convert an error to an RPC error.
+pub trait ToRpcError: core::error::Error + Send + Sync + 'static {
+    /// Converts the error to a JSON-RPC error object.
+    fn to_rpc_error(&self) -> jsonrpsee_types::ErrorObject<'static>;
+}
+
+impl ToRpcError for jsonrpsee_types::ErrorObject<'static> {
+    fn to_rpc_error(&self) -> jsonrpsee_types::ErrorObject<'static> {
+        self.clone()
+    }
+}
 
 /// Result alias
 pub type EthResult<T> = Result<T, EthApiError>;
