--- crates/rpc/rpc-types/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "reth-rpc-types"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-description = "Reth RPC types"
-
-[lints]
-workspace = true
-
-[dependencies]
-# misc
-jsonrpsee-types = { workspace = true, optional = true }
-
-[dev-dependencies]
-
-[features]
-default = ["jsonrpsee-types"]
-jsonrpsee-types = [
-    "dep:jsonrpsee-types",
-]
