--- crates/rpc/rpc-types/src/eth/mod.rs
@@ -1,3 +1,0 @@
-//! Ethereum related types
-
-pub(crate) mod error;
