+++ examples/custom-rlpx-subprotocol/Cargo.toml
@@ -15,7 +15,6 @@ reth-network-api.workspace = true
 reth-node-ethereum.workspace = true
 reth-provider = { workspace = true, features = ["test-utils"] }
 reth-primitives.workspace = true
-reth-rpc-types.workspace = true
 reth.workspace = true
 tokio-stream.workspace = true
 eyre.workspace = true
