+++ examples/db-access/Cargo.toml
@@ -11,7 +11,6 @@ reth-chainspec.workspace = true
 reth-db.workspace = true
 reth-primitives.workspace = true
 reth-provider.workspace = true
-reth-rpc-types.workspace = true
 reth-node-ethereum.workspace = true
 reth-node-types.workspace = true
